- Size map according to what's inside data-map-height. - Move resizeMapContainer to utils
