+++ projects/plugins/jetpack/extensions/blocks/map/mapbox-utils/index.js
@@ -26,20 +26,6 @@ export function fitMapToBounds( map, bounds ) {
 	} );
 }
 
-export function resizeMapContainer( container, fixedHeight = null ) {
-	if ( fixedHeight ) {
-		container.style.height = fixedHeight + 'px';
-	} else {
-		const blockWidth = container.offsetWidth;
-		const maxHeight =
-			window.location.search.indexOf( 'map-block-counter' ) > -1
-				? window.innerHeight
-				: window.innerHeight * 0.8;
-		const blockHeight = Math.min( blockWidth * ( 3 / 4 ), maxHeight );
-		container.style.height = blockHeight + 'px';
-	}
-}
-
 export function setMarkerHTML( el, markerColor ) {
 	el.innerHTML = `
 		<?xml version="1.0" encoding="UTF-8"?>
